perf(rust): complete the f32 migration — tuner, recovery, wire protocol
Final f32 pass: the cross-core wire protocol, the auto-tuner, and the
crash-recovery decision. Collapses the f32<->f64 cast boundary the earlier
rounds left at the controller edges.

- protocol: Status / Command / TuningSnapshot are now f32 EXCEPT the
  wall-clock fields, which stay wide: Status.timestamp (f64) and the
  scheduler start_time / seconds_until_start (u64) — f32 ulp at ~1.7e9 is
  128 s, which would wreck the delayed-start scheduler.
- tuner: hybrid f32 like pid/state — temps, SSR%, durations and the plateau
  window are f32; the absolute monotonic-seconds anchors (start_time,
  target_reached_time, last_plateau_check, the now params) stay f64 and only
  their small deltas narrow to f32.
- recovery: LastLogEntry / RecoveryDecision / check_recovery are f32
  (elapsed is a duration, no wall clock).
- KilnError and ControllerConfig follow, which removes the casts round 2
  had to add inside state.rs.
- controller: with everything f32, build_status / handle_command / the
  tuning bridge drop their per-field casts; only wall_s (timestamp) and the
  scheduler's u64 start_time stay wide. params.rs and ssr_schedule::new f32.

Consumers (kiln-app config/csv/recovery_io, kiln-sim, firmware platform)
cast f64-parsed config/log values at the boundary; KilnConfig stays f64 so
the persisted config JSON round-trips byte-identically, and
write_recovery_event_row keeps its f64 timestamp with the rest f32. Golden
replay tolerances relaxed for the new f32 paths; discrete tuner/recovery
outputs still match the f64 reference exactly.

Measured (stable build, real binary) vs the round-2 baseline:
  RAM (.data+.bss): 203,396 -> 202,460 B   (-936; control_task pool -744)
  Flash:            747,912 -> 747,032 B   (-880)

Cumulative vs pre-f32 baseline: RAM -4,520 B, flash -3,672 B, control_task
pool 6,624 -> 4,040 B (-39%).

+++ rust/kiln-app/src/config.rs
@@ -264,12 +264,12 @@ impl KilnConfig {
     /// The safety/rate/stall subset, as the `kiln-core` controller wants it.
     pub fn controller_config(&self) -> ControllerConfig {
         ControllerConfig {
-            max_temp: self.max_temp,
-            rate_measurement_window: self.rate_measurement_window,
-            rate_recording_interval: self.rate_recording_interval,
-            stall_check_interval: self.stall_check_interval,
+            max_temp: self.max_temp as f32,
+            rate_measurement_window: self.rate_measurement_window as f32,
+            rate_recording_interval: self.rate_recording_interval as f32,
+            stall_check_interval: self.stall_check_interval as f32,
             stall_consecutive_fails: self.stall_consecutive_fails,
-            stall_min_step_time: self.stall_min_step_time,
+            stall_min_step_time: self.stall_min_step_time as f32,
         }
     }
